+++ src/popup/Components/UpdateTranslations/types.ts
